fix(firmware): make the ESP32-S3 build configure and compile; gate it in CI
The production firmware has never linked, and no PR or push CI has ever
compiled it: `idf.py build` appeared only in release-artifacts.yml
(`release: published`) and firmware-perf-record.yml (needs a
`[self-hosted, esp32]` runner its own header says was never provisioned).
Five independently fatal build-system defects and four silently-excluded
components had accumulated behind that gap.

Build-system fixes (none of them application logic):

1. webui declared `PRIV_REQUIRES esp_vfs_fat`, which is a header directory,
   not an IDF component -- an unconditional configure failure. Corrected to
   `fatfs`. server.c includes no FAT/VFS header today; noted in place rather
   than dropped, since that is the webui author's call.
2. sdkconfig.defaults sets CONFIG_PARTITION_TABLE_CUSTOM_FILENAME=
   "partitions.csv" and no such file existed. Added, with nvs + phy_init +
   a single 3 MB factory app (no OTA anywhere in this firmware).
3. `components/hal` shadowed ESP-IDF's core `hal` component, so hal/aes_types.h
   and friends vanished and mbedtls could not compile. Renamed to
   `temper_hal`; all in-repo references updated.
4. temper_hal/CMakeLists.txt selected REQUIRES inside `if(CONFIG_IDF_TARGET)`,
   which is unset during IDF's requirements-expansion pass, so the ESP32
   sources compiled with no driver/esp_timer include paths. Keyed on
   ESP_PLATFORM, which idf.py sets on the cmake command line. Added the
   `esp_adc` requirement the ADC sources actually need.
5. Compile errors: IRAM_ATTR with no esp_attr.h in scope (hal_gpio_esp32.c,
   pll_control.c), ESP_LOG* with no esp_log.h (thermal_mass.c), and ~30
   -Werror=format sites passing int32_t HAL handles to %d (int32_t is
   `long int` on xtensa). All log-only; cast to (int).

Added the missing CMakeLists.txt for `control` and `safety`. This does NOT
link simulation code: safety.c's static simulated read_dc_bus_current()
returning sim_state.dc_bus_current, and the pll_sim_*/zvs_sim_* entry points,
all sit inside `#ifndef ESP_PLATFORM`, which IDF preprocesses away. Verified
with nm, not assumed -- no sim_state or *_sim_* symbol is defined anywhere in
the built objects, and the four protection reads appear as undefined (U).
`sensors` and `testing` still have no CMakeLists.txt, deliberately and with
reasons recorded: sensors' sources are already compiled into main (giving it
a component would create a main<->sensors cycle and duplicate symbols), and
testing is a header-only orphan with zero consumers.

The build now configures and compiles cleanly and fails at the link with 34
distinct undefined symbols / 141 undefined references. That is the honest
result and it is the deliverable: there is no sensor/peripheral layer, and
read_dc_bus_current and read_heatsink_temperature gate the IGBT
short-circuit, over-current and thermal trips. NONE of the missing symbols
is implemented, stubbed or weak-symboled here, and no source was removed to
make the build pass.

CI: new non-blocking job "Firmware Build (ESP32-S3 link ratchet)" in
firmware-tests.yml, on the same pinned espressif/idf:release-v5.3 image as
release-artifacts.yml. scripts/check_firmware_link.py prints the symbol and
reference counts to the job summary on every run, then ratchets against
.firmware-link-inventory on the .hash-order-inventory model: a new symbol,
a fixed-but-unrecorded symbol, any failure before the linker, a component
silently dropped for want of a CMakeLists.txt, or a simulation symbol
defined in the image all fail. It is green on a correct tree, so it has no
reason to get switched off. The job comment states exactly what would make
it blocking.

+++ firmware/components/temper_hal/esp32/hal_pwm_esp32.c
@@ -70,7 +70,7 @@ static hal_status_t esp32_pwm_init(hal_pwm_channel_t channel, const hal_pwm_conf
     }
     
     if (s_channels[channel].initialized) {
-        ESP_LOGW(TAG, "PWM channel %d already initialized", channel);
+        ESP_LOGW(TAG, "PWM channel %d already initialized", (int)channel);
         return HAL_ERROR_BUSY;
     }
     
@@ -209,7 +209,7 @@ static hal_status_t esp32_pwm_init(hal_pwm_channel_t channel, const hal_pwm_conf
     ch->initialized = true;
     
     ESP_LOGI(TAG, "PWM channel %d initialized: %lu Hz, %.1f%% duty, %u ns dead-time",
-             channel, (unsigned long)config->frequency_hz, config->duty_percent, config->dead_time_ns);
+             (int)channel, (unsigned long)config->frequency_hz, config->duty_percent, config->dead_time_ns);
     return HAL_OK;
 }
 
@@ -335,7 +335,7 @@ static hal_status_t esp32_pwm_start(hal_pwm_channel_t channel)
     }
     
     ch->state.running = true;
-    ESP_LOGI(TAG, "PWM channel %d started", channel);
+    ESP_LOGI(TAG, "PWM channel %d started", (int)channel);
     return HAL_OK;
 }
 
@@ -363,7 +363,7 @@ static hal_status_t esp32_pwm_stop(hal_pwm_channel_t channel)
     }
     
     ch->state.running = false;
-    ESP_LOGI(TAG, "PWM channel %d stopped", channel);
+    ESP_LOGI(TAG, "PWM channel %d stopped", (int)channel);
     return HAL_OK;
 }
 
@@ -455,7 +455,7 @@ static hal_status_t esp32_pwm_deinit(hal_pwm_channel_t channel)
     
     memset(ch, 0, sizeof(pwm_channel_t));
     
-    ESP_LOGI(TAG, "PWM channel %d deinitialized", channel);
+    ESP_LOGI(TAG, "PWM channel %d deinitialized", (int)channel);
     return HAL_OK;
 }
